requirements.txt
asyncpg==0.28.0
fastapi==0.101.0
python-dotenv==1.0.0
tortoise-orm==0.19.3
uvicorn==0.23.2